+++ CHANGELOG.md
@@ -9,8 +9,9 @@
 
 - [x] Add python3doc
 - [x] Fix english language
+- [x] Clean code for +/1
+- [ ] Display pause mode
 - [ ] Use constants variables 
-- [ ] Clean code for +/1
 
 # 1.2.1